# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "seekastyle@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/login": {
            "post": {
                "description": "user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "使用者登錄",
                "parameters": [
                    {
                        "description": "必填",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1-ok 0-fail",
                        "schema": {
                            "$ref": "#/definitions/handler.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/change-password": {
            "put": {
                "description": "update user login password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改使用者密碼",
                "parameters": [
                    {
                        "description": "必填",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserChangePwdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "get user info data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "查詢使用者資訊",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "add user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "使用者註冊",
                "parameters": [
                    {
                        "description": "必填",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.Data"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handler.Data": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "abcd.abcd.abcd"
                }
            }
        },
        "handler.GetResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.UserInfoResponse"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handler.UserChangePwdRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
        "handler.UserInfoResponse": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.UserLoginRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handler.UserRegisterRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}